refactoring to use hooks at the top
